Refactor : Exception Handle Class & Refactor : Orders Add API

+++ BE/tmong/src/main/java/com/teamonion/tmong/menu/MenuService.java
@@ -1,9 +1,11 @@
 package com.teamonion.tmong.menu;
 
-import com.teamonion.tmong.exception.CustomException;
-import com.teamonion.tmong.exception.CustomExceptionType;
-import com.teamonion.tmong.exception.ValidCustomException;
-import com.teamonion.tmong.exception.ValidExceptionType;
+import com.teamonion.tmong.exception.HandleRuntimeException;
+import com.teamonion.tmong.exception.GlobalExceptionType;
+import com.teamonion.tmong.exception.UnauthorizedException;
+import com.teamonion.tmong.member.MemberRole;
+import com.teamonion.tmong.security.JwtComponent;
+import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
@@ -13,6 +15,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.validation.constraints.NotNull;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Path;
@@ -21,6 +24,7 @@ import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
 
+@RequiredArgsConstructor
 @Service
 public class MenuService {
     private static Logger log = LoggerFactory.getLogger(MenuController.class);
@@ -28,20 +32,21 @@ public class MenuService {
     @Value("${download-path}")
     private String DOWNLOAD_PATH;
 
+    @NotNull
     private final MenuRepository menuRepository;
 
-    public MenuService(MenuRepository menuRepository) {
-        this.menuRepository = menuRepository;
-    }
-
+    @NotNull
+    private final JwtComponent jwtComponent;
 
     Long add(MenuSaveDto menuSaveDto) {
+        checkAdmin();
         MultipartFile imageFile = menuSaveDto.getImageFile();
 
         if (imageFile.getOriginalFilename().isEmpty()) {
-            throw new CustomException(CustomExceptionType.MENUIMAGE_NOT_FOUND);
+            throw new HandleRuntimeException(GlobalExceptionType.MENUIMAGE_NOT_FOUND);
         }
-        menuSaveDto.setImagePath(saveMenuImage(imageFile));
+
+        menuSaveDto.setImagePath(setMenuImagePath(imageFile));
 
         return menuRepository.save(menuSaveDto.toEntity()).getId();
     }
@@ -52,46 +57,45 @@ public class MenuService {
 
     @Transactional
     public void updateMenu(Long id, MenuSaveDto menuSaveDto) {
+        checkAdmin();
         Menu menu = menuRepository.findById(id)
-                .orElseThrow(() -> new ValidCustomException(ValidExceptionType.MENU_NOT_FOUND));
+                .orElseThrow(() -> new HandleRuntimeException(GlobalExceptionType.MENU_NOT_FOUND));
 
         String path = menu.getImagePath();
         MultipartFile imageFile = menuSaveDto.getImageFile();
 
         if (imageFile.getOriginalFilename().isEmpty()) {
-            throw new CustomException(CustomExceptionType.MENUIMAGE_NOT_FOUND);
+            throw new HandleRuntimeException(GlobalExceptionType.MENUIMAGE_NOT_FOUND);
         }
-        menuSaveDto.setImagePath(saveMenuImage(menuSaveDto.getImageFile()));
+        menuSaveDto.setImagePath(setMenuImagePath(menuSaveDto.getImageFile()));
         menu.update(menuSaveDto);
         menuRepository.save(menu);
 
         deleteMenuImage(path);
     }
 
-    public String saveMenuImage(MultipartFile imageFile) {
+    public String setMenuImagePath(MultipartFile imageFile) {
         try {
             int randomString = (int) (Math.random() * 10000) + 1;
             String fileName = System.currentTimeMillis() + "_" + randomString + "_" + imageFile.getOriginalFilename();
-
             String date = LocalDate.now().format(DateTimeFormatter.BASIC_ISO_DATE);
 
             File file = new File(DOWNLOAD_PATH + date + "/");
-
             file.mkdirs();
 
             Path path = Paths.get(DOWNLOAD_PATH + date + "/" + fileName);
-
             imageFile.transferTo(path);
 
             return DOWNLOAD_PATH + "/" + fileName;
         } catch (IOException e) {
             //e.getStackTrace();
-            throw new CustomException(CustomExceptionType.MENUIMAGE_RENDER_ERROR);
+            throw new HandleRuntimeException(GlobalExceptionType.MENUIMAGE_RENDER_ERROR);
         }
     }
 
     void deleteByMenuId(Long id) {
-        Menu menu = menuRepository.findById(id).orElseThrow(() -> new ValidCustomException(ValidExceptionType.MENU_NOT_FOUND));
+        checkAdmin();
+        Menu menu = menuRepository.findById(id).orElseThrow(() -> new HandleRuntimeException(GlobalExceptionType.MENU_NOT_FOUND));
         String path = menu.getImagePath();
 
         menuRepository.delete(menu);
@@ -108,5 +112,11 @@ public class MenuService {
 
     List<Menu> selectByName(String name) {
         return menuRepository.findByName(name);
+    }
+
+    private void checkAdmin() {
+        if (!jwtComponent.getClaimValueByToken(JwtComponent.ROLE).equals(MemberRole.ADMIN)) {
+            throw new UnauthorizedException();
+        }
     }
 }
